feat: modulos , componetizando y estilado
