+++ app.json
@@ -26,5 +26,8 @@
     "web": {
       "favicon": "./assets/favicon.png"
     }
+  },
+  "packagerOpts": {
+    "sourceExts": ["jsx", "js", "ts", "tsx"] 
   }
 }
