+++ index.js
@@ -1,0 +1,5 @@
+import { registerRootComponent } from "expo";
+
+import App from './src/index'
+
+registerRootComponent(App)
