+++ package.json
@@ -1,7 +1,6 @@
 {
   "name": "my-first-app",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
