+++ src/components/events/listItem/styles.js
@@ -1,0 +1,10 @@
+import { StyleSheet } from 'react-native';
+import {colors} from '../../../constants/colors'
+
+export const styles = StyleSheet.create({
+    listContainer: {
+        flex: 1,
+        backgroundColor: colors.backgroundListColor,
+        marginVertical: 20,
+    },
+});
