+++ src/components/index.js
@@ -1,0 +1,4 @@
+export {default as Input} from './input/index'
+export {default as Item} from './events/item/index'
+export {default as ListItem} from './events/listItem/index'
+export {default as DeleteModal} from './modal/index'
