+++ src/constants/colors.js
@@ -1,0 +1,8 @@
+export const colors = {
+    primary:'#52528C',
+    backgroundItemColor: '#52528C',
+    textItemColor: '#fff',
+    backgroundListColor:'#fff',
+    inputBorderColor:'violet',
+    inputColor:'violet',
+}
